Move code around a bit.
Move structs away from header.rs and utils.rs, keeping them closer to where they are used.
