+++ src/decoder.rs
@@ -1,4 +1,4 @@
-use crate::{header::Header, *};
+use crate::*;
 use bytes::{Buf, BytesMut, IntoBuf};
 
 /// Decode bytes from a [BytesMut] buffer as a [Packet] enum.
@@ -81,6 +81,44 @@ fn read_header(buffer: &mut BytesMut) -> Result<Option<(Header, usize)>, Error> 
     Err(Error::InvalidHeader)
 }
 
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(crate) struct Header {
+    pub typ: PacketType,
+    pub dup: bool,
+    pub qos: QoS,
+    pub retain: bool,
+}
+impl Header {
+    pub fn new(hd: u8) -> Result<Header, Error> {
+        let (typ, flags_ok) = match hd >> 4 {
+            1 => (PacketType::Connect, hd & 0b1111 == 0),
+            2 => (PacketType::Connack, hd & 0b1111 == 0),
+            3 => (PacketType::Publish, true),
+            4 => (PacketType::Puback, hd & 0b1111 == 0),
+            5 => (PacketType::Pubrec, hd & 0b1111 == 0),
+            6 => (PacketType::Pubrel, hd & 0b1111 == 0b0010),
+            7 => (PacketType::Pubcomp, hd & 0b1111 == 0),
+            8 => (PacketType::Subscribe, hd & 0b1111 == 0b0010),
+            9 => (PacketType::Suback, hd & 0b1111 == 0),
+            10 => (PacketType::Unsubscribe, hd & 0b1111 == 0b0010),
+            11 => (PacketType::Unsuback, hd & 0b1111 == 0),
+            12 => (PacketType::Pingreq, hd & 0b1111 == 0),
+            13 => (PacketType::Pingresp, hd & 0b1111 == 0),
+            14 => (PacketType::Disconnect, hd & 0b1111 == 0),
+            _ => (PacketType::Connect, false),
+        };
+        if !flags_ok {
+            return Err(Error::InvalidHeader);
+        }
+        Ok(Header {
+            typ,
+            dup: hd & 0b1000 != 0,
+            qos: QoS::from_u8((hd & 0b110) >> 1)?,
+            retain: hd & 1 == 1,
+        })
+    }
+}
+
 pub(crate) fn read_string(buffer: &mut BytesMut) -> Result<String, Error> {
     String::from_utf8(read_bytes(buffer)?).map_err(|e| Error::InvalidString(e.utf8_error()))
 }
@@ -96,7 +134,7 @@ pub(crate) fn read_bytes(buffer: &mut BytesMut) -> Result<Vec<u8>, Error> {
 
 #[cfg(test)]
 mod test {
-    use crate::{decoder::read_header, header::Header, *};
+    use crate::decoder::*;
     use bytes::BytesMut;
 
     macro_rules! header {
