+++ src/utils.rs
@@ -192,56 +192,3 @@ impl QosPid {
         }
     }
 }
-
-/// Message that the server should publish when the client disconnects.
-///
-/// Sent by the client in the [Connect] packet. [MQTT 3.1.3.3].
-///
-/// [Connect]: struct.Connect.html
-/// [MQTT 3.1.3.3]: http://docs.oasis-open.org/mqtt/mqtt/v3.1.1/os/mqtt-v3.1.1-os.html#_Toc398718031
-#[derive(Debug, Clone, PartialEq)]
-pub struct LastWill {
-    pub topic: String,
-    pub message: Vec<u8>,
-    pub qos: QoS,
-    pub retain: bool,
-}
-
-/// Sucess value of a [Connack] packet.
-///
-/// See [MQTT 3.2.2.3] for interpretations.
-///
-/// [Connack]: struct.Connack.html
-/// [MQTT 3.2.2.3]: http://docs.oasis-open.org/mqtt/mqtt/v3.1.1/os/mqtt-v3.1.1-os.html#_Toc398718035
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum ConnectReturnCode {
-    Accepted,
-    RefusedProtocolVersion,
-    RefusedIdentifierRejected,
-    ServerUnavailable,
-    BadUsernamePassword,
-    NotAuthorized,
-}
-impl ConnectReturnCode {
-    pub(crate) fn to_u8(&self) -> u8 {
-        match *self {
-            ConnectReturnCode::Accepted => 0,
-            ConnectReturnCode::RefusedProtocolVersion => 1,
-            ConnectReturnCode::RefusedIdentifierRejected => 2,
-            ConnectReturnCode::ServerUnavailable => 3,
-            ConnectReturnCode::BadUsernamePassword => 4,
-            ConnectReturnCode::NotAuthorized => 5,
-        }
-    }
-    pub(crate) fn from_u8(byte: u8) -> Result<ConnectReturnCode, Error> {
-        match byte {
-            0 => Ok(ConnectReturnCode::Accepted),
-            1 => Ok(ConnectReturnCode::RefusedProtocolVersion),
-            2 => Ok(ConnectReturnCode::RefusedIdentifierRejected),
-            3 => Ok(ConnectReturnCode::ServerUnavailable),
-            4 => Ok(ConnectReturnCode::BadUsernamePassword),
-            5 => Ok(ConnectReturnCode::NotAuthorized),
-            n => Err(Error::InvalidConnectReturnCode(n)),
-        }
-    }
-}
